chore(CNAME): deletes `CNAME` file
This file is not needed with hosting at Cloudflare Pages.

The `CNAME` file in the context of a static website, especially when deployed via GitHub Pages, serves a crucial role in domain mapping. Essentially, it tells GitHub Pages the custom domain that you intend to use for your site. When you configure your website to be accessible via a custom domain, you place this domain name inside a `CNAME` file at the root of your published site directory. This is a directive used by GitHub Pages to correctly route requests for your custom domain to your GitHub-hosted site.